refactor: rename solving to reaction
