+++ expertsystem/amplitude/canonical_decay.py
@@ -3,11 +3,11 @@
 from typing import Any, Callable, List, Optional
 
 from expertsystem.particles import Spin
-from expertsystem.solving.quantum_numbers import (
+from expertsystem.reaction.quantum_numbers import (
     InteractionProperties,
     ParticleWithSpin,
 )
-from expertsystem.solving.topology import StateTransitionGraph
+from expertsystem.reaction.topology import StateTransitionGraph
 
 from .helicity_decay import (
     HelicityAmplitudeGenerator,
