+++ expertsystem/amplitude/helicity_decay.py
@@ -3,11 +3,11 @@ import logging
 from typing import Dict, List, Optional, Tuple, Union
 
 from expertsystem.particles import ParticleCollection, Spin
-from expertsystem.solving.combinatorics import (
+from expertsystem.reaction.combinatorics import (
     perform_external_edge_identical_particle_combinatorics,
 )
-from expertsystem.solving.quantum_numbers import ParticleWithSpin
-from expertsystem.solving.topology import StateTransitionGraph, Topology
+from expertsystem.reaction.quantum_numbers import ParticleWithSpin
+from expertsystem.reaction.topology import StateTransitionGraph, Topology
 
 from .model import (
     AmplitudeModel,
