+++ expertsystem/reaction/combinatorics.py
