+++ tests/unit/solving/test_combinatorics.py
@@ -6,7 +6,7 @@ from math import factorial
 import pytest
 
 from expertsystem import io
-from expertsystem.solving.combinatorics import (
+from expertsystem.reaction.combinatorics import (
     KinematicRepresentation,
     _generate_kinematic_permutations,
     _generate_outer_edge_permutations,
@@ -15,7 +15,7 @@ from expertsystem.solving.combinatorics import (
     get_kinematic_representation,
     initialize_graph,
 )
-from expertsystem.solving.topology import (
+from expertsystem.reaction.topology import (
     InteractionNode,
     SimpleStateTransitionTopologyBuilder,
     StateTransitionGraph,
